added little feature as spring boot ;)
