+++ szkoczki/src/main/java/pl/pp/skoczki/SkoczkiLogic/game/pawn/Color.java
@@ -1,4 +1,4 @@
-package pl.pp.skoczki.SkoczkiLogic;
+package pl.pp.skoczki.SkoczkiLogic.game.pawn;
 
 
 public enum Color {
